Review

+++ src/main/java/com/project/phonecaseshop/repository/DesignRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface DesignRepository extends JpaRepository<Design, Long> {
+public interface DesignRepository extends JpaRepository<Design, Integer> {
     List<Design> findByProductId_ProductId(int productId);
     void deleteByProductId_ProductId(int productId);
 }
